% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_interval_censored_like.R
\name{weibull_interval_loglike}
\alias{weibull_interval_loglike}
\title{Create interval-censored Weibull log-likelihood function
For use in joint model survival component}
\usage{
weibull_interval_loglike(survObject)
}
\description{
Create interval-censored Weibull log-likelihood function
For use in joint model survival component
}
